my_app/requirements.txt
fastapi[all]
pandas
joblib
gcsfs
scikit-learn
lightgbm
xgboost
catboost
google-cloud-bigquery
openpyxl
mlflow
